Remove legacy content verification result model
