+++ .github/scripts/content-verification/findings-gate-cli.ts
@@ -1,6 +1,6 @@
-import { parseVerificationManifest } from "./agentic-gate.ts";
 import { readContentVerificationArtifacts } from "./artifacts.ts";
 import { reduceFindingEvents } from "./finding-events.ts";
+import { parseVerificationManifest } from "./manifest-validation.ts";
 import { parseVerificationScope } from "./scope.ts";
 
 function required(name: string): string {
