+++ .github/workflows/README.md
@@ -5,10 +5,10 @@ This repository uses gh-aw for four Agentic GitHub Actions tasks. A shared runti
 The shared runtime adoption is recorded in
 [ADR 0001](../../docs/adr/0001-use-gh-aw-for-agentic-github-workflows.md).
 [ADR 0003](../../docs/adr/0003-use-mutable-finding-events-for-content-verification.md)
-records the current findings architecture;
+records the current findings architecture and supersedes
 [ADR 0002](../../docs/adr/0002-resolve-inconclusive-content-verification-through-trusted-issues.md)
-retains the legacy contract's migration history until final cleanup. The
-general state and trust invariants are maintained as
+without rewriting that record's historical context. The general state and trust
+invariants are maintained as
 [Agent run-state Knowledge](../../knowledge/github-actions/agent-run-state-and-reruns.md).
 
 ## Workflow inventory
@@ -62,7 +62,7 @@ Repository content under review, issue and pull-request text, and external pages
 Each scheduled task also supports manual dispatch and keeps its own name, schedule, scope, and concurrency identity. Before inference, a trusted step derives an immutable target manifest from `git ls-files`, the parsed [Knowledge index](../../knowledge/index.md), and the checked-out revision. Target discovery is implemented in [the content-verification scripts](../scripts/content-verification/targets.ts).
 
 Task-specific sources own what to analyze. All three import the
-[mutable-findings contract](shared/content-verification-findings.md).
+[shared content-verification contract](shared/content-verification.md).
 
 A Knowledge target owns one leaf. A Skill target owns its `SKILL.md` and tracked files below the same directory. Package-level references and `.github/workflows/shared/*.md` components are independent shared-reference targets. Each otherwise unowned `AGENTS.md`, root `CONTEXT.md`, and file under `docs/agents/` is an instruction target. Each root `.github/workflows/*.md` source except this README is an Agentic workflow target, and each `.github/scripts/*/prompts/` directory is one prompt target. An invalid index, mutable revision, empty scope, duplicate target, or duplicate file ownership fails before the Agent runs.
